% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0001_ontology.R
\name{read_acm0001_waste_type_shapes}
\alias{read_acm0001_waste_type_shapes}
\title{Load ACM0001 waste type applicability shape}
\usage{
read_acm0001_waste_type_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shaclR:read_shacl]{shaclR::read_shacl()}}.
}
\description{
Load ACM0001 waste type applicability shape
}
